+++ conf/script/src/build_system/cmd/setup/cli/target_info.py
